AnswerLotto객체/유효성 검사 (리팩토링)
